Work around lag caused by help-echo text
Emacs likes to call the help-echo function every time we move the mouse cursor,
even if only moving it a single pixel. Since it takes a few milliseconds to get
the title of the linked item, we need to cache the result and only update it
when we're actually hovering over a different link.